+++ src/components/VideoPlayer/VideoPlayer.test.tsx
@@ -1,86 +1,82 @@
-// import { expect } from "chai";
-// import * as Enzyme from "enzyme";
-// import * as React from "react";
-// import VideoPlayer from "./VideoPlayer";
+import { expect } from "chai";
+import * as Enzyme from "enzyme";
+import * as React from "react";
+import VideoPlayer from "./VideoPlayer";
+import { VideoPlayerAspectRatios, VideoPlayerTypes } from "./VideoPlayerTypes";
 
-/*
-describe("HorizontalRule: check for basic element", () => {
+describe("VideoPlayer", () => {
   let container;
-  before(() => {
-    container = Enzyme.mount(<HorizontalRule />);
+
+  describe("YouTube player", () => {
+    before(() => {
+      container = Enzyme.mount(
+        <VideoPlayer
+          videoType={VideoPlayerTypes.YouTube}
+          videoId="nm-dD2tx6bk"
+        />
+      );
+    });
+
+    it("Renders VideoPlayer container", () => {
+      expect(container.find(".video-player").exists()).to.equal(true);
+    });
+
+    it("Renders VideoPlayer iframe", () => {
+      expect(container.find("iframe").exists()).to.equal(true);
+    });
+
+    it("Renders YouTube video", () => {
+      expect(container.find("iframe").prop("src")).to.contain("youtube.com");
+    });
   });
 
-  it("Renders HorizontalRule component", () => {
-    expect(container.find("hr").exists()).to.equal(true);
+  describe("Vimeo player", () => {
+    before(() => {
+      container = Enzyme.mount(
+        <VideoPlayer videoType={VideoPlayerTypes.Vimeo} videoId="474719268" />
+      );
+    });
+
+    it("Renders VideoPlayer container", () => {
+      expect(container.find(".video-player").exists()).to.equal(true);
+    });
+
+    it("Renders VideoPlayer iframe", () => {
+      expect(container.find("iframe").exists()).to.equal(true);
+    });
+
+    it("Renders Vimeo video", () => {
+      expect(container.find("iframe").prop("src")).to.contain("vimeo.com");
+    });
+  });
+
+  describe("aspect ratio", () => {
+    it("Renders with 4:3 aspect ratio", () => {
+      container = Enzyme.mount(
+        <VideoPlayer
+          videoType={VideoPlayerTypes.Vimeo}
+          videoId="474719268"
+          aspectRatio={VideoPlayerAspectRatios.FourByThree}
+        />
+      );
+
+      expect(
+        container.find(".video-player").hasClass("video-player--four-by-three")
+      ).to.equal(true);
+    });
+
+    it("Renders with 1:1 aspect ratio", () => {
+      container = Enzyme.mount(
+        <VideoPlayer
+          videoType={VideoPlayerTypes.Vimeo}
+          videoId="474719268"
+          aspectRatio={VideoPlayerAspectRatios.Square}
+        />
+      );
+
+      expect(
+        container.find(".video-player").hasClass("video-player--square")
+      ).to.equal(true);
+    });
   });
 });
-
-describe("HorizontalRule: check for custom class", () => {
-  let container;
-  before(() => {
-    container = Enzyme.mount(<HorizontalRule className={"custom-hr"} />);
-  });
-
-  it("Renders HorizontalRule component", () => {
-    expect(container.find("hr").exists()).to.equal(true);
-  });
-
-  it("Renders with proper custom class", () => {
-    expect(container.find("hr").hasClass("custom-hr")).to.equal(true);
-  });
-});
-
-describe("HorizontalRule: check for custom dimensions", () => {
-  let container;
-  before(() => {
-    container = Enzyme.mount(
-      <HorizontalRule className={"custom-hr"} width={"720px"} height={"5rem"} />
-    );
-  });
-
-  it("Renders HorizontalRule component", () => {
-    expect(container.find("hr").exists()).to.equal(true);
-  });
-
-  it("Renders with correct width and height values", () => {
-    expect(container.find("hr").get(0).props.style).to.have.property(
-      "width",
-      "720px"
-    );
-    expect(container.find("hr").get(0).props.style).to.have.property(
-      "height",
-      "5rem"
-    );
-  });
-});
-
-describe("HorizontalRule: check for custom width and custom alignment", () => {
-  let container;
-  before(() => {
-    container = Enzyme.mount(
-      <HorizontalRule className={"custom-hr"} width={"80%"} align={"left"} />
-    );
-  });
-
-  it("Renders HorizontalRule component", () => {
-    expect(container.find("hr").exists()).to.equal(true);
-  });
-
-  it("Renders with correct width and height values", () => {
-    expect(container.find("hr").get(0).props.style).to.have.property(
-      "width",
-      "80%"
-    );
-    expect(container.find("hr").get(0).props.style).to.have.property(
-      "height",
-      "2px"
-    );
-  });
-
-  it("Renders with correct alignment", () => {
-    expect(container.find("hr").hasClass("horizontal-rule--left")).to.equal(
-      true
-    );
-  });
-});
-*/