+++ CHANGELOG.md
@@ -15,7 +15,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Changes
 
-- Updates stories for `List` component to use MDX format.
+- Updates stories for the `List` component to use MDX format.
+- Updates stories for the `Pagination` component to use MDX format.
 - Changes `DisplaySizes` to `HeadingDisplaySizes`
 - Changes `Button` to use `16px` for left/right padding for buttons with text; icon only buttons still use `8px` left/right padding