+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates how children are rendered in the `List` component. Direct children `li`, `dt`, and `dd` elements can still be passed, but specific data structures can be passed to the `listItems` prop to render the appropriate children element.
 - Changes `DisplaySizes` to `HeadingDisplaySizes`
 - Changes `Button` to use `16px` for left/right padding for buttons with text; icon only buttons still use `8px` left/right padding
+- Updates to the contributing documentation, pull request template, and the README.
 
 ## 0.23.1