+++ CHANGELOG.md
@@ -10,8 +10,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds `Notification` component.
+- Adds the ability to render a `textarea` element from the `TextInput` component.
 - Adds Style Guide `Buttons`
-- Adds `Notification` component
 
 ### Changes