+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Changes `DisplaySizes` to `HeadingDisplaySizes`
 - Changes `Button` to use `16px` for left/right padding for buttons with text; icon only buttons still use `8px` left/right padding
 - Updates to the contributing documentation, pull request template, and the README.
+- Changes `TextInput` label text to `font-weight: 500`.
 
 ## 0.23.1