+++ CHANGELOG.md
@@ -20,9 +20,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates stories for the `List` component to use MDX format.
 - Updates stories for the `Pagination` component to use MDX format.
 - Updates how children are rendered in the `List` component. Direct children `li`, `dt`, and `dd` elements can still be passed, but specific data structures can be passed to the `listItems` prop to render the appropriate children element.
-- Changes `DisplaySizes` to `HeadingDisplaySizes`
-- Changes `Button` to use `16px` for left/right padding for buttons with text; icon only buttons still use `8px` left/right padding
 - Updates to the contributing documentation, pull request template, and the README.
+- Changes `DisplaySizes` to `HeadingDisplaySizes`.
+- Changes `Skeleton Loader` to be configurable via props.
+- Changes `Button` to use `16px` for left/right padding for buttons with text; icon only buttons still use `8px` left/right padding.
 - Changes `TextInput` label text to `font-weight: 500`.
 
 ## 0.23.1