+++ CHANGELOG.md
@@ -12,8 +12,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Adds `Form` component.
 - Adds `Notification` component.
-- Adds the ability to render a `textarea` element from the `TextInput` component.
 - Adds Style Guide `Buttons`.
+- Adds `VideoPlayer` component.
+- Adds the ability to render a `textarea` element from the `TextInput` component.
 
 ### Changes